Remove the code attempting to set GRANGE_FLAG_REFPOS_HAVE_* fields.
These are used internally in tg_iface_g.c to determine whether we need
to store the Position, Size or ID fields via a simplistic delta
mechanism.  Outside of that file they should never be set or queried.

This fix removes a warning about the in-memory and on-disc
representations differing, which is bogus as these bit fields are
zeroed after reading anyway.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden/trunk@3700 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1