Work around for a bug where some bins are not covered by child bins. Ie:
:-----------------------------------: Parent

                      :-------------: Left
           :------------------------: Right

I don't know what causes this yet, but it makes
calculate_consensus_simple sometiems find very large bins (the parent)
on which to compute the cached consensus, taking a lot of time.  For
now we simply reject an attempt to compute a very long cached
consensus as it is probably not the write thing to do.

(This really needs a total rewrite.)


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden/trunk@3753 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1