haplo_hash bug fixes.
In cache_dup, ensure that haplo_hash and haplo_timestamp are initialized
when duplicating a contig.  This prevents a dangling pointer from being left
in the parent if haplo_hash is freed in the child.

In cache_flush, ensure that haplo_hash is freed from the parent where
necessary when flushing a contig block.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden/trunk@3757 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1