Fix rounding error in qplot_redraw which triggered an assertion failure.
Inverted the xzoom factor (xz) so that the x coordinate in the plot can
be worked out by multiplying by xz instead of dividing.  This seems to
fix a rounding error that could make the x position go out of bounds
when zoomed out a long way.  Multiplying should also be slightly faster.

In a belt-and-braces approach, also allocated the xs array one bigger
than strictly necessary so any remaining rounding errors shouldn't cause
a problem.  The assertions on x are adjusted to reflect this.

Only one of the xp and xs arrays is used, depending on the value of xz.  So
now only the one that is needed gets allocated.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden/trunk@3778 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1